+++ tests/conftest.py
@@ -8,10 +8,15 @@ import yaml
 from jinja2 import Environment, FileSystemLoader
 from conda_build.utils import copy_into
 
-from conda_smithy.configure_feedstock import conda_forge_content, _load_forge_config
+from conda_smithy.configure_feedstock import (
+    conda_forge_content,
+    _load_forge_config,
+)
 
 
-RecipeConfigPair = collections.namedtuple("RecipeConfigPair", ("recipe", "config"))
+RecipeConfigPair = collections.namedtuple(
+    "RecipeConfigPair", ("recipe", "config")
+)
 
 
 @pytest.fixture(scope="function")
@@ -46,15 +51,18 @@ def testing_workdir(tmpdir, request):
 def config_yaml(testing_workdir):
     config = {"python": ["2.7", "3.5"], "r_base": ["3.3.2", "3.4.2"]}
     os.makedirs(os.path.join(testing_workdir, "recipe"))
-    os.makedirs(os.path.join(testing_workdir, "migrations"))
     with open(os.path.join(testing_workdir, "config.yaml"), "w") as f:
         f.write("docker:\n")
         f.write("  fallback_image:\n")
         f.write("  - centos:6\n")
-    with open(os.path.join(testing_workdir, "recipe", "default_config.yaml"), "w") as f:
+    with open(
+        os.path.join(testing_workdir, "recipe", "default_config.yaml"), "w"
+    ) as f:
         yaml.dump(config, f, default_flow_style=False)
         # need selectors, so write these more manually
-        f.write(dedent("""\
+        f.write(
+            dedent(
+                """\
         target_platform: 
         - win-64        # [win]
         - win-32        # [win]
@@ -68,17 +76,24 @@ def config_yaml(testing_workdir):
         - c_compiler    # [win]
         - python        # [win]
         - vc            # [win]
-        """))
+        """
+            )
+        )
     # dummy file that needs to be present for circle ci.  This is created by the init function
     os.makedirs(os.path.join(testing_workdir, ".circleci"))
     with open(
-        os.path.join(testing_workdir, ".circleci", "checkout_merge_commit.sh"), "w"
+        os.path.join(testing_workdir, ".circleci", "checkout_merge_commit.sh"),
+        "w",
     ) as f:
         f.write("echo dummy file")
-    with open(os.path.join(testing_workdir, "recipe", "short_config.yaml"), "w") as f:
+    with open(
+        os.path.join(testing_workdir, "recipe", "short_config.yaml"), "w"
+    ) as f:
         config = {"python": ["2.7"]}
         yaml.dump(config, f, default_flow_style=False)
-    with open(os.path.join(testing_workdir, "recipe", "long_config.yaml"), "w") as f:
+    with open(
+        os.path.join(testing_workdir, "recipe", "long_config.yaml"), "w"
+    ) as f:
         config = {"python": ["2.7", "3.5", "3.6"]}
         yaml.dump(config, f, default_flow_style=False)
     with open(os.path.join(testing_workdir, "conda-forge.yml"), "w") as f:
@@ -107,8 +122,10 @@ requirements:
     return RecipeConfigPair(
         str(config_yaml),
         _load_forge_config(
-            config_yaml, exclusive_config_file=os.path.join(
-                config_yaml, "recipe", "default_config.yaml")
+            config_yaml,
+            exclusive_config_file=os.path.join(
+                config_yaml, "recipe", "default_config.yaml"
+            ),
         ),
     )
 
@@ -133,8 +150,10 @@ requirements:
     return RecipeConfigPair(
         str(config_yaml),
         _load_forge_config(
-            config_yaml, exclusive_config_file=os.path.join(
-                config_yaml, "recipe", "default_config.yaml")
+            config_yaml,
+            exclusive_config_file=os.path.join(
+                config_yaml, "recipe", "default_config.yaml"
+            ),
         ),
     )
 
@@ -161,8 +180,10 @@ about:
     return RecipeConfigPair(
         str(config_yaml),
         _load_forge_config(
-            config_yaml, exclusive_config_file=os.path.join(
-                config_yaml, "recipe", "default_config.yaml")
+            config_yaml,
+            exclusive_config_file=os.path.join(
+                config_yaml, "recipe", "default_config.yaml"
+            ),
         ),
     )
 
@@ -189,8 +210,8 @@ about:
     return RecipeConfigPair(
         str(config_yaml),
         _load_forge_config(
-            config_yaml, exclusive_config_file=os.path.join(
-                config_yaml, "conda-forge.yml")
+            config_yaml,
+            exclusive_config_file=os.path.join(config_yaml, "conda-forge.yml"),
         ),
     )
 
@@ -215,17 +236,27 @@ about:
     """
         )
 
-    with open(os.path.join(config_yaml, "migrations", "zlib.yaml"), "w") as fh:
-        fh.write("""
+    os.makedirs(
+        os.path.join(config_yaml, ".ci_support", "migrations"), exist_ok=True
+    )
+    with open(
+        os.path.join(config_yaml, ".ci_support", "migrations", "zlib.yaml"),
+        "w",
+    ) as fh:
+        fh.write(
+            """
 zlib:
     - 1000
-""")
+"""
+        )
 
     return RecipeConfigPair(
         str(config_yaml),
         _load_forge_config(
-            config_yaml, exclusive_config_file=os.path.join(
-                config_yaml, "recipe", "default_config.yaml")
+            config_yaml,
+            exclusive_config_file=os.path.join(
+                config_yaml, "recipe", "default_config.yaml"
+            ),
         ),
     )
 
@@ -234,25 +265,48 @@ zlib:
 def recipe_migration_cfep9_downgrade(config_yaml, recipe_migration_cfep9):
     # write a downgrade migrator that lives next to the current migrator.
     # Only this, more recent migrator should apply.
-    with open(os.path.join(config_yaml, "migrations", "zlib-downgrade.yaml"), "w") as fh:
-        fh.write("""
+    os.makedirs(
+        os.path.join(config_yaml, ".ci_support", "migrations"), exist_ok=True
+    )
+    with open(
+        os.path.join(
+            config_yaml, ".ci_support", "migrations", "zlib-downgrade.yaml"
+        ),
+        "w",
+    ) as fh:
+        fh.write(
+            """
 migration_ts: 1.0
 zlib:
     - 999
-""")
-    #return recipe_migration_cfep9
+"""
+        )
+    # return recipe_migration_cfep9
     return RecipeConfigPair(
         str(config_yaml),
         _load_forge_config(
-            config_yaml, exclusive_config_file=os.path.join(
-                config_yaml, "recipe", "default_config.yaml")
+            config_yaml,
+            exclusive_config_file=os.path.join(
+                config_yaml, "recipe", "default_config.yaml"
+            ),
         ),
     )
 
+
 @pytest.fixture(scope="function")
 def recipe_migration_win_compiled(config_yaml, py_recipe):
-    with open(os.path.join(config_yaml, "migrations", "vc-migrate.yaml"), "w") as fh:
-        fh.write(dedent("""
+    os.makedirs(
+        os.path.join(config_yaml, ".ci_support", "migrations"), exist_ok=True
+    )
+    with open(
+        os.path.join(
+            config_yaml, ".ci_support", "migrations", "vc-migrate.yaml"
+        ),
+        "w",
+    ) as fh:
+        fh.write(
+            dedent(
+                """
         migration_ts: 1.0
         c_compiler:    # [win]
             - vs2008   # [win]
@@ -268,12 +322,16 @@ def recipe_migration_win_compiled(config_yaml, py_recipe):
               - c_compiler      # [win]
               - cxx_compiler    # [win]
               - vc              # [win]
-        """))
+        """
+            )
+        )
     return RecipeConfigPair(
         str(config_yaml),
         _load_forge_config(
-            config_yaml, exclusive_config_file=os.path.join(
-                config_yaml, "recipe", "default_config.yaml")
+            config_yaml,
+            exclusive_config_file=os.path.join(
+                config_yaml, "recipe", "default_config.yaml"
+            ),
         ),
     )
 
@@ -300,8 +358,10 @@ about:
     return RecipeConfigPair(
         str(config_yaml),
         _load_forge_config(
-            config_yaml, exclusive_config_file=os.path.join(
-                config_yaml, "recipe", "default_config.yaml")
+            config_yaml,
+            exclusive_config_file=os.path.join(
+                config_yaml, "recipe", "default_config.yaml"
+            ),
         ),
     )
 
@@ -328,8 +388,10 @@ about:
     return RecipeConfigPair(
         str(config_yaml),
         _load_forge_config(
-            config_yaml, exclusive_config_file=os.path.join(
-                config_yaml, "recipe", "default_config.yaml")
+            config_yaml,
+            exclusive_config_file=os.path.join(
+                config_yaml, "recipe", "default_config.yaml"
+            ),
         ),
     )
 
@@ -354,8 +416,10 @@ about:
     return RecipeConfigPair(
         str(config_yaml),
         _load_forge_config(
-            config_yaml, exclusive_config_file=os.path.join(
-                config_yaml, "recipe", "default_config.yaml")
+            config_yaml,
+            exclusive_config_file=os.path.join(
+                config_yaml, "recipe", "default_config.yaml"
+            ),
         ),
     )
 
@@ -390,8 +454,10 @@ skip_render:
     return RecipeConfigPair(
         str(config_yaml),
         _load_forge_config(
-            config_yaml, exclusive_config_file=os.path.join(
-                config_yaml, "recipe", "default_config.yaml")
+            config_yaml,
+            exclusive_config_file=os.path.join(
+                config_yaml, "recipe", "default_config.yaml"
+            ),
         ),
     )